+++ CHANGES.rst
@@ -5,6 +5,16 @@ Changelog
 - new MINOR version for added functionality in a backwards compatible manner
 - new PATCH version for backwards compatible bug fixes
 
+v1.1.0
+--------
+2021-12-27:
+    - add Store Api DELETE method
+    - add Store Api GET method
+    - add Store Api GET LIST method
+    - add Store Api PATCH method
+    - add Store Api PUT method
+
+
 v1.0.0
 --------
 2021-12-26: initial release